+++ docs/release_notes.rst
@@ -1,17 +1,19 @@
 Release notes
 ==============
 
-0.1.2 (2025-11-11)
+0.1.2 (2025-11-17)
 -------------
 * Fixed documentation of command-line options
 * Added MouseEyetracker emulator, for testing without Tobii hardware
 * Fixed timestamp synchronization, now timestamps are raw time from core of Psychopy
-* Fixed dependency issues. Psychopy version is now pinned to 2024.1.4 to avoid compatibility issues.
+* Fixed dependency issues. Psychopy version is not pinned to 2024.1.4, but troubleshooting section added to cope with possible problems with new version.
 * Fixed issues with mouse gaze data collection (multiple gaze points when no RBS button clicked)
 * Refactored code for analyzers and data loaders.
 * Added full set of readings from eyetracker (not limited to average pupil size, or average gaze point)
 * Added HeatMapAnalyzer, FixationAnalyzer, FocusMapAnalyzer, SaccadeAnalyzer complete code
 * Fixed issues with pyglet requirements on Windows
+* Moved bounding box detection and stimulus drawing to Dataset class.
+* Removed enable_mode -- it is now handled by optional configuration field.
 
 
 0.1.1 (2025-10-28)